pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oioioi"
version = "0.2.0.dev"
description = "The web frontend of the SIO2 Project contesting system"
authors = [
    {name = "The SIO2 Team", email = "sio2@sio2project.mimuw.edu.pl"}
]
requires-python = ">=3.11"
dependencies = [
    "Django>=5.2,<5.3", # Szkopul is now compatible with Django 6.0
    "pytz>=2023.3,<2023.4",
    "SQLAlchemy>=2.0,<2.1.0",
    "beautifulsoup4>=4.12,<4.13",
    "PyYAML>=6.0.3,<6.1",
    "python-dateutil>=2.8,<2.9",
    "django-two-factor-auth==1.18.0",
    "django-otp==1.2.2",
    "django-registration-redux>=2.12,<2.13",
    "Celery>=5.5,<5.6",
    "drf-spectacular>=0.28.0",
    "dj-pagination>=2.5,<2.6",
    "django-compressor>=4.6,<4.7",
    "Pygments>=2.19,<2.20",
    "django-libsass>=0.9,<0.10",
    "django-debug-toolbar<6.0.0", # 6.1.0 causes an obscure test failure.
    "django-extensions>=4.1,<4.2",
    "djangorestframework>=3.14,<3.15",
    "django-oauth-toolkit>=3.2,<3.3",
    "Werkzeug",
    "setuptools>=61.0",
    "requests>=2.32,<2.33",
    "fpdf>=1.7,<1.8",
    "unicodecsv>=0.14,<0.15",
    "dnslib>=0.9,<0.10",
    "bleach>=6.3,<6.4",
    "chardet>=5.1",
    "django-gravatar2>=1.4,<1.5",
    "django-mptt>=0.18,<0.19",
    "mistune<2.0", # 2.0 is breaking
    "pika>=1.3,<1.4",
    "Unidecode>=1.3,<1.4",
    "sentry-sdk>=2.16.0,<2.17.0",
    "fontawesomefree>=6.4,<6.5",
    # A library allowing to nest inlines in django admin.
    # Used in quizzes module for adding new quizzes.
    "django-nested-admin>=4.1,<4.2",
    # Library for parsing dates and timedelta
    "humanize<=4.9.0",
    # SIO2 dependencies:
    "filetracker>=2.3.1,<3.0",
    "django-simple-captcha>=0.6.1,<0.7",
    "phonenumbers>=8.13,<8.14",
    "pdfminer.six==20260107",
    # https://stackoverflow.com/questions/73929564/entrypoints-object-has-no-attribute-get-digital-ocean
    "importlib-metadata==4.13.0",
    "supervisor==4.3.0", # previously http://github.com/Supervisor/supervisor/zipball/master#egg=supervisor==4.0.0.dev0
    "django-supervisor@git+https://github.com/sio2project/django-supervisor#egg=django-supervisor", # previously http://github.com/badochov/djsupervisor/zipball/master#egg=djsupervisor==0.4.0
    "cffi==2.0.0",
    "django-appconf>=1.2,<1.3",
    "django-phonenumber-field>=8.4,<8.5",
    "cryptography==46.0.5",
    "legacy-cgi",
    "sioworkers@https://github.com/sio2project/sioworkers/archive/refs/tags/v1.5.9.tar.gz",
    "pylibmc",
    "psycopg2-binary",
    "aiohttp>=3.12.2,<3.13.0",
    "aio-pika>=9.5.5,<9.6.0",
    "cachetools>=6.0.0",
    "websockets>=16.0,<16.1",
]

[dependency-groups]
dev = [
    "ruff==0.13.2",
    "elastic-apm",
    "python-memcached",
    "pre-commit==4.2.0",
    "tox>=4.0.0",
    "tox-uv>=1.0.0",
    "inquirer>=3.4.1",
    "pytest>=7.2,<8.0",
    "pytest-cov>=7.0,<7.1",
    "pytest-django>=4.11,<5.0",
    "pytest-html>=4.2,<4.3",
    "pytest-metadata>=3.0,<4.0",
    "pytest-xdist>=3.8,<3.9",
    "coverage>=7.4.0",
    "coverage[toml]>=7.4.0",
]
docs = [
    "sphinx>7.1.0",
    "furo",
    "sphinx-autoapi>3.5.0",
    "pydantic",
    "pydantic-settings",
]

[project.urls]
Homepage = "https://github.com/sio2project"

[project.scripts]
oioioi-create-config = "oioioi.deployment.create_config:main"

[project.entry-points."sio.compilers"]
"g++14_2_cpp23_amd64" = "oioioi.sioworkers.compilers:run_cpp_gcc14_2_cpp23_amd64"

[tool.setuptools]
packages = {find = {exclude = ["ez_setup"]}}
include-package-data = true

[tool.ruff]
line-length = 160
lint.select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "B", # flake8-bugbear
    "UP", # pyupgrade
    "C4" # flake8-comprehensions
]
lint.ignore = [
    "B904",
    "B026",
    "C416",
]
src = ["oioioi"]
exclude = [
    "migrations",
    "oioioi/**/*settings.py",
    "extra/sample-configs/runnable-test-settings.py",
    "oioioi/problems/tests/__init__.py"
]
target-version = "py313"
